requirements.txt
asgiref==3.6.0
Brotli==1.0.9
crispy-bootstrap4==2022.1
Django==4.2.1
django-crispy-forms==2.0
django-jalali==6.0.1
django-jalali-date==1.0.2
django-render-partial==0.4
django-table-sort==0.5.0
django-widget-tweaks==1.4.12
fonttools==4.40.0
jdatetime==4.1.1
Pillow==9.5.0
psycopg==3.1.9
pycparser==2.21
reportlab==4.0.4
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
webencodings==0.5.1
zopfli==0.2.2
